Add macOS and Windows (MSYS2) support to the C/C++ build
The makefile, simulator/estimator sources, and docs only ever targeted
Linux. Add Homebrew-based macOS support (OpenMP via libomp, Eigen/Boost
path detection) and MSYS2/MinGW-w64 support for Windows, plus CI to
build on all three platforms:

- makefile: OS-detect via uname -s, add Homebrew include/lib paths for
  libomp/eigen/boost on Darwin, bump to c++14 (current Homebrew Eigen
  requires it)
- guard <malloc.h> (absent on macOS) in common_header.h, FN/Def.h,
  ML/Def.h
- fmath.hpp's exp() fast path is x86 SSE2-only; gate it behind an x86
  arch check in common_header.h and fall back to std exp() elsewhere
  (fixes a hard compile failure on Apple Silicon)
- fix narrowing-conversion errors in npy header calls that Clang treats
  as hard errors (HH/Initialization.h, Causality/main.cpp)
- include/mkdir.h: branch mkdir() vs Windows' _mkdir(), and rename the
  local helper to make_dir_recursive to avoid colliding with the
  Windows CRT's own _mkdir
- add .github/workflows/build.yml building on ubuntu-latest,
  macos-latest, and windows-latest (MSYS2/MinGW64)
- fix .gitignore's blanket *.yml rule swallowing the new CI workflow
- update README.md/CLAUDE.md with per-OS build instructions
- trim CLAUDE.md duplication/derivable content

Verified end-to-end on macOS (Apple Silicon): make all -j builds all
10 binaries, simHH runs a full simulation, and calCausality computes
causality from real spike data. Windows is unverified locally and
relies on the new CI job.

+++ HH/Initialization.h
@@ -174,7 +174,7 @@ void Record_connect_matrix()
 
 		string output_filename = oss.str();
 		ofstream fp = save_npy_header<double>(
-			output_filename, {N, N});
+			output_filename, {(size_t)N, (size_t)N});
 		for (int i = 0; i < N; i++)
 			fp.write(reinterpret_cast<const char*>(Connect_Matrix[i]), N * sizeof(double));
 		fp.close();
@@ -187,7 +187,7 @@ void Record_connect_matrix()
 			if (random_S != 0) {
 				output_filename.erase(output_filename.end() - 4, output_filename.end());
 				output_filename.append("_strength.npy");
-				ofstream fp = save_npy_header<double>(output_filename, {N, N});
+				ofstream fp = save_npy_header<double>(output_filename, {(size_t)N, (size_t)N});
 				for (int i = 0; i < N; i++)
 					fp.write(reinterpret_cast<const char*>(CS[i]), N * sizeof(double));
 				fp.close();
